+++ src/core/modules/insert/tests.rs
@@ -2,7 +2,7 @@ use input::{Event, KeyCode};
 use view::assert_rendered_output;
 
 use super::*;
-use crate::{assert_process_result, module::testutil::module_test};
+use crate::{assert_process_result, core::testutil::module_test};
 
 #[test]
 fn activate() {
