+++ src/core/testutil/assert_process_result.rs
@@ -1,7 +1,7 @@
 use anyhow::Error;
 use input::Event;
 
-use crate::module::{ExitStatus, ProcessResult, State};
+use crate::core::module::{ExitStatus, ProcessResult, State};
 
 fn format_process_result(
 	event: Option<Event>,
@@ -104,16 +104,16 @@ pub fn _assert_process_result(
 #[macro_export]
 macro_rules! assert_process_result {
 	($actual:expr) => {
-		crate::module::testutil::_assert_process_result(&$actual, None, None, None, &None, &None)
+		crate::core::testutil::_assert_process_result(&$actual, None, None, None, &None, &None)
 	};
 	($actual:expr, error = $error:expr, exit_status = $exit_status:expr) => {
-		crate::module::testutil::_assert_process_result(&$actual, None, None, Some($exit_status), &Some($error), &None)
+		crate::core::testutil::_assert_process_result(&$actual, None, None, Some($exit_status), &Some($error), &None)
 	};
 	($actual:expr, state = $state:expr) => {
-		crate::module::testutil::_assert_process_result(&$actual, None, Some($state), None, &None, &None)
+		crate::core::testutil::_assert_process_result(&$actual, None, Some($state), None, &None, &None)
 	};
 	($actual:expr, state = $state:expr, external_command = $external_command:expr) => {
-		crate::module::testutil::_assert_process_result(
+		crate::core::testutil::_assert_process_result(
 			&$actual,
 			None,
 			Some($state),
@@ -123,19 +123,19 @@ macro_rules! assert_process_result {
 		)
 	};
 	($actual:expr, state = $state:expr, error = $error:expr) => {
-		crate::module::testutil::_assert_process_result(&$actual, None, Some($state), None, &Some($error), &None)
+		crate::core::testutil::_assert_process_result(&$actual, None, Some($state), None, &Some($error), &None)
 	};
 	($actual:expr, event = $event:expr) => {
-		crate::module::testutil::_assert_process_result(&$actual, Some($event), None, None, &None, &None)
+		crate::core::testutil::_assert_process_result(&$actual, Some($event), None, None, &None, &None)
 	};
 	($actual:expr, event = $event:expr, state = $state:expr) => {
-		crate::module::testutil::_assert_process_result(&$actual, Some($event), Some($state), None, &None, &None)
+		crate::core::testutil::_assert_process_result(&$actual, Some($event), Some($state), None, &None, &None)
 	};
 	($actual:expr, event = $event:expr, exit_status = $exit_status:expr) => {
-		crate::module::testutil::_assert_process_result(&$actual, Some($event), None, Some($exit_status), &None, &None)
+		crate::core::testutil::_assert_process_result(&$actual, Some($event), None, Some($exit_status), &None, &None)
 	};
 	($actual:expr, event = $event:expr, external_command = $external_command:expr) => {
-		crate::module::testutil::_assert_process_result(
+		crate::core::testutil::_assert_process_result(
 			&$actual,
 			Some($event),
 			None,
@@ -146,6 +146,6 @@ macro_rules! assert_process_result {
 	};
 
 	($actual:expr, external_command = $external_command:expr) => {
-		crate::module::testutil::_assert_process_result(&$actual, None, None, None, &None, &Some($external_command))
+		crate::core::testutil::_assert_process_result(&$actual, None, None, None, &None, &Some($external_command))
 	};
 }
